+++ lib/migration_generator/operation.ex
@@ -45,7 +45,7 @@ defmodule AshPostgres.MigrationGenerator.Operation do
         }) do
       [
         "add #{inspect(attribute.name)}",
-        "references(#{inspect(table)}",
+        "references(:#{table}",
         [
           "type: #{inspect(attribute.type)}",
           "column: #{inspect(destination_field)}",
@@ -71,11 +71,10 @@ defmodule AshPostgres.MigrationGenerator.Operation do
         }) do
       [
         "add #{inspect(attribute.name)}",
-        "references(#{inspect(table)}",
+        "references(:#{table}",
         [
           "type: #{inspect(attribute.type)}",
           "column: #{inspect(destination_field)}",
-          "name: \"\#\{prefix()\}_#{table}_#{attribute.name}_fkey\"",
           "prefix: \"public\""
         ],
         ")",
@@ -118,11 +117,10 @@ defmodule AshPostgres.MigrationGenerator.Operation do
         }) do
       [
         "add #{inspect(attribute.name)}",
-        "references(#{inspect(table)}",
+        "references(:#{table}",
         [
           "type: #{inspect(attribute.type)}",
-          "column: #{inspect(destination_field)}",
-          "name: \"\#\{prefix()\}_#{table}_#{attribute.name}_fkey\""
+          "column: #{inspect(destination_field)}"
         ],
         ")",
         maybe_add_default(attribute.default),
@@ -137,7 +135,7 @@ defmodule AshPostgres.MigrationGenerator.Operation do
         }) do
       [
         "add #{inspect(attribute.name)}",
-        "references(#{inspect(table)}",
+        "references(:#{table}",
         [
           "type: #{inspect(attribute.type)}",
           "column: #{inspect(destination_field)}"
@@ -233,16 +231,13 @@ defmodule AshPostgres.MigrationGenerator.Operation do
 
     defp reference(%{strategy: :context}, %{
            type: type,
-           name: name,
            references: %{
              multitenancy: %{strategy: :context},
              table: table,
              destination_field: destination_field
            }
          }) do
-      "references(#{inspect(table)}, type: #{inspect(type)}, column: #{inspect(destination_field)}, name: \"\#\{prefix()\}_#{
-        table
-      }_#{name}_fkey\")"
+      "references(:#{table}, type: #{inspect(type)}, column: #{inspect(destination_field)})"
     end
 
     defp reference(%{strategy: :attribute, attribute: source_attribute}, %{
@@ -253,7 +248,7 @@ defmodule AshPostgres.MigrationGenerator.Operation do
              destination_field: destination_field
            }
          }) do
-      "references(#{inspect(table)}, type: #{inspect(type)}, column: #{inspect(destination_field)}, with: [#{
+      "references(:#{table}, type: #{inspect(type)}, column: #{inspect(destination_field)}, with: [#{
         source_attribute
       }: :#{destination_attribute}])"
     end
@@ -268,7 +263,7 @@ defmodule AshPostgres.MigrationGenerator.Operation do
              }
            }
          ) do
-      "references(#{inspect(table)}, type: #{inspect(type)}, column: #{inspect(destination_field)}, prefix(): \"public\")"
+      "references(:#{table}, type: #{inspect(type)}, column: #{inspect(destination_field)}, prefix: \"public\")"
     end
 
     defp reference(
@@ -281,7 +276,7 @@ defmodule AshPostgres.MigrationGenerator.Operation do
              }
            }
          ) do
-      "references(#{inspect(table)}, type: #{inspect(type)}, column: #{inspect(destination_field)})"
+      "references(:#{table}, type: #{inspect(type)}, column: #{inspect(destination_field)})"
     end
 
     def down(op) do
@@ -302,22 +297,14 @@ defmodule AshPostgres.MigrationGenerator.Operation do
     # We only need to drop it before altering an attribute with `references/3`
     defstruct [:attribute, :table, :multitenancy, :direction, no_phase: true]
 
-    def up(%{attribute: attribute, table: table, multitenancy: multitenancy, direction: :up}) do
-      if multitenancy && multitenancy.strategy == :context do
-        "drop constraint(:#{table}, \"\#\{prefix()\}_#{table}_#{attribute.name}_fkey\")"
-      else
-        "drop constraint(:#{table}, \"#{table}_#{attribute.name}_fkey\")"
-      end
+    def up(%{attribute: attribute, table: table, direction: :up}) do
+      "drop constraint(:#{table}, \"#{table}_#{attribute.name}_fkey\")"
     end
 
     def up(_), do: ""
 
-    def down(%{attribute: attribute, table: table, multitenancy: multitenancy, direction: :down}) do
-      if multitenancy && multitenancy.strategy == :context do
-        "drop constraint(:#{table}, \"\#\{prefix()\}_#{table}_#{attribute.name}_fkey\")"
-      else
-        "drop constraint(:#{table}, \"#{table}_#{attribute.name}_fkey\")"
-      end
+    def down(%{attribute: attribute, table: table, direction: :down}) do
+      "drop constraint(:#{table}, \"#{table}_#{attribute.name}_fkey\")"
     end
 
     def down(_), do: ""