+++ test/support/api.ex
@@ -6,5 +6,6 @@ defmodule AshPostgres.Test.Api do
     resource(AshPostgres.Test.Post)
     resource(AshPostgres.Test.Comment)
     resource(AshPostgres.Test.IntegerPost)
+    resource(AshPostgres.Test.Rating)
   end
 end