+++ test/support/resources/comment.ex
@@ -10,7 +10,12 @@ defmodule AshPostgres.Test.Comment do
 
   actions do
     read(:read)
-    create(:create)
+
+    create :create do
+      argument(:rating, :map)
+
+      change(manage_relationship(:rating, :ratings, on_missing: :ignore, on_match: :create))
+    end
   end
 
   attributes do
@@ -20,5 +25,10 @@ defmodule AshPostgres.Test.Comment do
 
   relationships do
     belongs_to(:post, AshPostgres.Test.Post)
+
+    has_many(:ratings, AshPostgres.Test.Rating,
+      destination_field: :resource_id,
+      context: %{data_layer: %{table: "comment_ratings"}}
+    )
   end
 end